file upload works!

+++ includes/functions.inc.php
@@ -324,12 +324,17 @@ function makeResource($ng, $name, $icon) {
 
     $filename = $_FILES['icon']['name'];
     $filetmpname = $_FILES['icon']['tmp_name'];
-    
-    $folder = 'img/resources/';
-    //function for saving the uploaded images in a specific folder
-    move_uploaded_file($filetmpname, $folder.$filename);
+        
+    $folder = '../img/resources/';
+    if (file_exists($folder.$filename) !==false) {
+        header("location: /NG/admanage.php?error=emptyinput");
+        exit();
+    }
+    else {
+        move_uploaded_file($filetmpname, $folder.$filename);
+    }
 
-    mysqli_stmt_bind_param($stmt, "ss", $name, $icon);
+    mysqli_stmt_bind_param($stmt, "ss", $name, $filename);
     mysqli_stmt_execute($stmt);
     mysqli_stmt_close($stmt);
 }
